requirements.txt
flask>=3.1.2
qrcode>=8.2
pillow>=12.0.0
gunicorn>=23.0.0
python-dotenv==1.2.1